+++ monocle/gerrit/review.py
@@ -31,9 +31,9 @@ class GerritCrawlerArgs(object):
     updated_since: str
     loop_delay: int
     command: str
-    index: str
     base_url: str
     repository: str
+    db: object
 
 
 class ReviewesFetcher(object):