+++ monocle/github/graphql.py
@@ -41,41 +41,40 @@ class GithubGraphQLQuery(object):
 
     log = logging.getLogger(__name__)
 
-    def __init__(self, token):
+    def __init__(self, token_getter):
         self.url = 'https://api.github.com/graphql'
-        self.headers = {
-            'Authorization': 'token %s' % token,
-            'User-Agent': 'change-metrics/monocle',
-        }
         self.session = requests.session()
         # Will get every 25 requests
         self.get_rate_limit_rate = 25
         self.query_count = 0
         # Set an initial value
         self.quota_remain = 5000
-        self.get_rate_limit()
-        self.retry_after = False
+        self.token_getter = token_getter
+
+    def get_token(self) -> str:
+        return self.token_getter.get_token()[0]
+
+    def get_headers(self) -> dict:
+        headers = {
+            'Authorization': 'token %s' % self.get_token(),
+            'User-Agent': 'change-metrics/monocle',
+        }
+        self.log.debug('request headers: %s' % headers)
+        return headers
 
     def get_rate_limit(self):
-        try:
-            ratelimit = self.getRateLimit()
-        except requests.exceptions.ConnectionError:
-            sleep(5)
-            ratelimit = self.getRateLimit()
-        self.quota_remain = ratelimit['remaining']
-        self.resetat = datetime.strptime(ratelimit['resetAt'], '%Y-%m-%dT%H:%M:%SZ')
-        self.log.info(
-            "Got rate limit data: remain %s resetat %s"
-            % (self.quota_remain, self.resetat)
-        )
+        ratelimit = self.getRateLimit()
+        if ratelimit:
+            self.quota_remain = ratelimit['remaining']
+            self.resetat = datetime.strptime(ratelimit['resetAt'], '%Y-%m-%dT%H:%M:%SZ')
+            self.log.info(
+                "Got rate limit data: remain %s resetat %s"
+                % (self.quota_remain, self.resetat)
+            )
 
     # https://developer.github.com/v3/guides/best-practices-for-integrators/#dealing-with-abuse-rate-limits
     def wait_for_call(self):
-        if self.retry_after:
-            self.log.info("Waiting for %s secs (Retry-After)" % self.retry_after)
-            sleep(self.retry_after)
-            self.retry_after = False
-        elif self.quota_remain <= 150:
+        if self.quota_remain <= 150:
             until_reset = self.resetat - datetime.utcnow()
             self.log.info(
                 "Quota remain: %s/calls delay until "
@@ -97,11 +96,12 @@ class GithubGraphQLQuery(object):
           }
         }'''
         data = self.query(qdata, skip_get_rate_limit=True)
-        try:
-            return data['data']['rateLimit']
-        except KeyError:
-            self.log.error('No rate limit data: %s' % data)
-            raise RequestException('No rate limit data: %s' % data)
+        if data:
+            try:
+                return data['data']['rateLimit']
+            except KeyError:
+                self.log.error('No rate limit data: %s' % data)
+                raise RequestException('No rate limit data: %s' % data)
 
     @retry(
         after=after_log(log, logging.INFO),
@@ -112,27 +112,30 @@ class GithubGraphQLQuery(object):
     )
     def query(self, qdata, skip_get_rate_limit=False):
         if not skip_get_rate_limit:
-            if (
-                not self.retry_after
-                and self.query_count % self.get_rate_limit_rate == 0
-            ):
+            if self.query_count % self.get_rate_limit_rate == 0:
                 self.get_rate_limit()
             self.wait_for_call()
         data = {'query': qdata}
         try:
             r = self.session.post(
-                url=self.url, json=data, headers=self.headers, timeout=30.3
+                url=self.url, json=data, headers=self.get_headers(), timeout=30.3
             )
-        except requests.exceptions.ConnectionError:
+        except (
+            requests.exceptions.ConnectionError,
+            requests.exceptions.ChunkedEncodingError,
+        ):
             raise RequestException("Error connecting to the API")
         self.query_count += 1
         if 'retry-after' in r.headers:
-            self.log.info('Got Retry-After: %s' % r.headers['retry-after'])
-            self.retry_after = int(r.headers['retry-after'])
+            self.log.info('Got Retry-After: %s, sleeping...' % r.headers['retry-after'])
+            sleep(int(r.headers['retry-after']))
         if not r.status_code != "200":
             self.log.error('No ok response code: %s' % r)
             raise RequestException("No ok response code: %s" % r.text)
         ret = r.json()
+        if 'Bad credentials' == ret.get('message', ''):
+            self.log.info('Query forbidden due to bad credentials')
+            ret = {}
         if 'errors' in ret:
             self.log.error("Errors in response: %s" % ret)
             if (
@@ -154,5 +157,11 @@ class GithubGraphQLQuery(object):
                     # the failing commit(s). So return the data to the caller and
                     # move on.
                     return ret
-            raise RequestException("Errors in response: %s" % ret['errors'])
+            is_forbiden = any([error['type'] == 'FORBIDDEN'] for error in ret['errors'])
+            if is_forbiden:
+                # Do not raise to not retrigger tenacity
+                self.log.info('Query forbidden due to unsuffcient token ACLs')
+                ret = {}
+            else:
+                raise RequestException("Errors in response: %s" % ret['errors'])
         return ret